CSP-256 more code coverage

+++ src/SFA.DAS.Roatp.Domain/Entities/NationalAchievementRateOverall.cs
@@ -14,7 +14,7 @@ namespace SFA.DAS.Roatp.Domain.Entities
             {
                 Age = source.Age,
                 SectorSubjectArea = source.SectorSubjectArea,
-                ApprenticeshipLevel = (int)source.ApprenticeshipLevel,
+                ApprenticeshipLevel = source.ApprenticeshipLevel,
                 OverallCohort = source.OverallCohort,
                 OverallAchievementRate = source.OverallAchievementRate,
             };
